requirements.txt

asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==23.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
